+++ src/github.com/getlantern/detour/mock.go
@@ -1,7 +1,6 @@
 package detour
 
 import (
-	"golang.org/x/net/ipv4"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -9,7 +8,7 @@ import (
 )
 
 var (
-	closeURL, timeOutURL, timeOut2ndTimeURL, echoURL, proxiedURL string
+	closeURL, timeoutURL, timeout2ndTimeURL, echoURL, proxiedURL string
 	directMsg                                                    string = "hello direct"
 	detourMsg                                                    string = "hello detour"
 )
@@ -18,12 +17,11 @@ func closeHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		hj := w.(http.Hijacker)
 		conn, _, _ := hj.Hijack()
-		ipv4.NewPacketConn(conn)
 		conn.Close()
 	}
 }
 
-func timeOutHandler(when int, d time.Duration) http.HandlerFunc {
+func timeoutHandler(when int, d time.Duration) http.HandlerFunc {
 	count := 0
 	return func(w http.ResponseWriter, r *http.Request) {
 		count = count + 1
@@ -47,14 +45,14 @@ var servers []*httptest.Server
 func startMockServers(t *testing.T) {
 	servers = []*httptest.Server{
 		httptest.NewServer(closeHandler()),
-		httptest.NewServer(timeOutHandler(1, 1*time.Second)),
-		httptest.NewServer(timeOutHandler(2, 1*time.Second)),
+		httptest.NewServer(timeoutHandler(1, 1*time.Second)),
+		httptest.NewServer(timeoutHandler(2, 1*time.Second)),
 		httptest.NewServer(echoHandler{directMsg}),
 		httptest.NewServer(echoHandler{detourMsg}),
 	}
 	closeURL = servers[0].URL
-	timeOutURL = servers[1].URL
-	timeOut2ndTimeURL = servers[2].URL
+	timeoutURL = servers[1].URL
+	timeout2ndTimeURL = servers[2].URL
 	echoURL = servers[3].URL
 	proxiedURL = servers[4].URL
 }
